Added random access get of an arcfile record.  Added primitive cmd-line processing and wrapper script.  Made underlying MappedByteBufferInputStream support random seeks (Its effectively a RandomAccessInputStream). * src/java/org/archive/io/GZIPMemberInputStream.java     Formatting.     (testGZIPheader): Refactoring so the code to test gzip member gets moved     out to a method so can test buffer made by me or made elsewhere.     (read(byte[] b, int off, int len)): Added. * src/java/org/archive/io/MappedByteBufferInputStream.java     (read(byte[] b, int off, int len)): Added.     (getPosition, setPosition): Added. * src/java/org/archive/io/arc/ARCReader.java     Javadoc. Formatting.     (get): Added implementation.     (main): Added. * src/java/org/archive/io/arc/ARCReaderFactory.java     (MAX_HEADER_LINE_LENGTH): Removed.  Unused.     Renamed NuArcReader to be ARCReader. * src/java/org/archive/io/arc/ARCRecord.java     Formatting. * src/java/org/archive/io/arc/ARCRecordMetaData.java     Made offset a long. * src/java/org/archive/io/arc/ARCWriterTest.java     Added testing of random get of arc record. * src/java/org/archive/io/Position.java src/scripts/arcreader     Wrapper for arcreader.

+++ heritrix/src/java/org/archive/io/arc/ARCRecordMetaData.java
@@ -69,6 +69,9 @@ public class ARCRecordMetaData
     /**
      * Test required field is present in hash.
      *
+     * @param fields Map of fields.
+     * @param requiredField Field to test for.
+     * 
      * @exception IOException If required field is not present.
      */
     protected void testRequiredField(Map fields, String requiredField)
@@ -127,8 +130,8 @@ public class ARCRecordMetaData
     /**
      * @return Offse into arcfile at which this record begins.
      */
-    public int getOffset() {
-        return ((Integer)this.headerFields.get(ABSOLUTE_OFFSET_KEY)).intValue();
+    public long getOffset() {
+        return ((Long)this.headerFields.get(ABSOLUTE_OFFSET_KEY)).longValue();
     }
 
     /**
